+++ filament/backend/src/metal/MetalEnums.mm
@@ -198,6 +198,16 @@ MTLPixelFormat getMetalFormat(MetalContext* context, TextureFormat format) noexc
 
             case TextureFormat::DXT1_RGB: return MTLPixelFormatInvalid;
             case TextureFormat::DXT1_SRGB: return MTLPixelFormatInvalid;
+
+            case TextureFormat::RED_RGTC1:              return MTLPixelFormatBC4_RUnorm;
+            case TextureFormat::SIGNED_RED_RGTC1:       return MTLPixelFormatBC4_RSnorm;
+            case TextureFormat::RED_GREEN_RGTC2:        return MTLPixelFormatBC5_RGUnorm;
+            case TextureFormat::SIGNED_RED_GREEN_RGTC2: return MTLPixelFormatBC5_RGSnorm;
+
+            case TextureFormat::RGB_BPTC_SIGNED_FLOAT:      return MTLPixelFormatBC6H_RGBFloat;
+            case TextureFormat::RGB_BPTC_UNSIGNED_FLOAT:    return MTLPixelFormatBC6H_RGBUfloat;
+            case TextureFormat::RGBA_BPTC_UNORM:            return MTLPixelFormatBC7_RGBAUnorm;
+            case TextureFormat::SRGB_ALPHA_BPTC_UNORM:      return MTLPixelFormatBC7_RGBAUnorm_sRGB;
             default: break;
         }
     }